pyproject.toml
[tool.poetry]
name = "sg-send-qa"
version = "v0.2.17"
description = "Browser automation test suite and living documentation generator for SG/Send"
authors = ["The Cyber Boardroom"]
readme = "README.md"
packages = [{include = "sg_send_qa"}]

[tool.poetry.dependencies]
python = "^3.12"
playwright = "^1.40"
fastapi = "^0.115"
uvicorn = "^0.34"
pillow = "^11.0"
osbot-utils = ">=3.72.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.25"
httpx = "^0.27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["."]
addopts = "--import-mode=importlib"
